In WeatherRepository I registered 2 search criteria in the database.
Created methods in WeatherService:
1. Search by city and country, wrote the logic of work.
2. Search by city, country and date of creation of the record, the logic of work is not written.
3. Saving the city weather in the database, information about the temperature in a given city is taken from 3 APIs, then the arithmetic average is calculated and this temperature value is entered into the database, with the values of the city and country.

Created a WeatherController with two methods:
1. GET request with City path variable, in order to send a request to 3 weather APIs, to form values - this is a temporary solution in the future I plan to implement a request function at a set time, while the application is running.
2. GET request with city and country path variable to display all records from the database according to the given criteria.

+++ src/main/java/com/myproject/weatherrestapi/entity/Weather.java
@@ -19,7 +19,7 @@ public class Weather {
     private String city;
 
     @Column(name = "temperature")
-    private Double temperature;
+    private Integer temperature;
 
     @Column(name = "created_at")
     private OffsetDateTime createdAt;
@@ -56,11 +56,11 @@ public class Weather {
         this.city = city;
     }
 
-    public Double getTemperature() {
+    public Integer getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(Double temperature) {
+    public void setTemperature(Integer temperature) {
         this.temperature = temperature;
     }
