+++ src/main/resources/db/changelog/script.sql
@@ -2,6 +2,6 @@ create table weather (
                          id serial primary key,
                          country varchar(255),
                          city varchar(255),
-                         temperature double precision,
+                         temperature int,
                          created_at timestamp
 )
